- bot context folder and name changed to telegram/types directory - create content messages writings fixed

+++ src/bots/telegram/middlewares/threadPostGuard.middleware.ts
@@ -1,30 +1,31 @@
 import { Context, MiddlewareFn } from "telegraf";
 import { getAdminsId } from "../utils/getAdminsId.util";
+import { publicTopicIds } from "../../../constants";
 
 export const threadPostGuard: MiddlewareFn<Context> = async (
-    ctx: Context,
-    next: () => Promise<void>
+  ctx: Context,
+  next: () => Promise<void>,
 ) => {
-    try{
-const msg = ctx.message;
+  try {
+    const msg = ctx.message;
     const chat = ctx.chat;
     const admins = await getAdminsId();
     const senderId = ctx.from?.id;
 
     if (
-        ["message", "command"].includes(ctx.updateType) &&
-        chat &&
-        chat?.type == "supergroup" &&
-        msg?.message_thread_id != undefined &&
-        senderId &&
-        admins.includes(senderId) == false
+      ["message", "command"].includes(ctx.updateType) &&
+      chat &&
+      chat?.type == "supergroup" &&
+      !publicTopicIds.includes(msg?.message_thread_id) &&
+      msg?.message_thread_id != undefined &&
+      senderId &&
+      admins.includes(senderId) == false
     ) {
-        ctx.deleteMessage(msg.message_id);
-        return;
+      ctx.deleteMessage(msg.message_id);
+      return;
     }
     await next();
-    }catch(e){
-      console.log("error happened ::::::")
-      return
-    }
+  } catch (e) {
+    return;
+  }
 };
